refactor: enable config based on active feature
