+++ runtime/extensions/Cargo.toml
@@ -13,6 +13,7 @@ publish = false
 targets = ["x86_64-unknown-linux-gnu"]
 
 [dependencies]
+cfg-if = "1.0"
 codec.workspace = true
 log.workspace = true
 
@@ -36,13 +37,11 @@ rand = "0.8.5"
 [features]
 default = ["std"]
 std = [
+ "log/std",
  "codec/std",
  "frame-support/std",
  "frame-system/std",
- "log/std",
- "pallet-assets/std",
  "pallet-contracts/std",
- "pallet-api/std",
  "pop-primitives/std",
  "sp-core/std",
  "sp-std/std",
@@ -53,5 +52,5 @@ runtime-benchmarks = [
  "pallet-assets/runtime-benchmarks",
  "pallet-contracts/runtime-benchmarks",
 ]
-pop-devnet = []
+pop-devnet = ["pallet-assets/std", "pallet-api/std"]
 pop-testnet = []
