+++ runtime/extensions/src/config/mod.rs
@@ -1,0 +1,4 @@
+#[cfg(feature = "pop-devnet")]
+pub(crate) mod devnet;
+#[cfg(feature = "pop-testnet")]
+pub(crate) mod testnet;
