+++ runtime/testnet/Cargo.toml
@@ -138,6 +138,8 @@ std = [
     "polkadot-parachain-primitives/std",
     "polkadot-runtime-common/std",
     "pop-primitives/std",
+    "pop-runtime-extensions/std",
+    "pop-runtime-extensions/pop-testnet",
     "scale-info/std",
     "sp-api/std",
     "sp-io/std",
